- Updated view models: refactored int id to long id

+++ shared/src/commonMain/kotlin/presentation/ui/main/cart/view_model/CartEvent.kt
@@ -5,9 +5,9 @@ import business.core.UIComponent
 
 sealed class CartEvent {
 
-    data class DeleteFromBasket(val id: Int) : CartEvent()
+    data class DeleteFromBasket(val id: Long) : CartEvent()
 
-    data class AddProduct(val id: Int) : CartEvent()
+    data class AddProduct(val id: Long) : CartEvent()
 
    data object OnRemoveHeadFromQueue : CartEvent()
