+++ shared/src/commonMain/kotlin/presentation/ui/main/comment/view_model/CommentViewModel.kt
@@ -63,7 +63,7 @@ class CommentViewModel(
     }
 
 
-    private fun onUpdateProductId(id: Int) {
+    private fun onUpdateProductId(id: Long) {
         state.value = state.value.copy(productId = id)
     }
